requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Environment and configuration
python-dotenv==1.0.0

# Payment processing
stripe==7.8.0

# Email
requests==2.31.0

# Data processing (for admin/data import only)
pandas==2.1.4
openpyxl==3.1.2

# Security enhancements
Flask-Talisman==1.1.0  # Security headers and CSP
Flask-WTF==1.2.1       # CSRF protection
cryptography==41.0.8   # Enhanced cryptography
Flask-Limiter==3.5.0   # Rate limiting

# Production server
gunicorn==21.2.0

# Monitoring and logging
sentry-sdk[flask]==1.38.0  # Error tracking (optional)

# Development dependencies (optional)
python-decouple==3.8